refactor(rating): rename RatingItemStep to RatimItemIndicator
