+++ packages/core/src/Rating/index.ts
@@ -1,3 +1,3 @@
 export { default as RatingItem, type RatingItemProps } from './RatingItem.vue'
-export { default as RatingItemStep } from './RatingItemStep.vue'
+export { default as RatingItemIndicator } from './RatingItemIndicator.vue'
 export { default as RatingRoot, type RatingRootProps } from './RatingRoot.vue'
